% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\name{DfUniqueByCols}
\alias{DfUniqueByCols}
\title{Reduce data frame by picking the first row of blocks for which \code{cols} has the same values}
\usage{
DfUniqueByCols(d, cols, drop = FALSE)
}
\description{
Reduce data frame by picking the first row of blocks for which \code{cols} has the same values
}
