% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RtestingHelpers.R
\name{getTests}
\alias{getTests}
\title{Legacy internal test function (getTests)}
\usage{
getTests(prefixes = list(c("rTest", "rExp"), c("T", "E")), which = -2)
}
\arguments{
\item{prefixes}{variable name prefixes for detecting test computations and expectations}

\item{which}{which `sys.frame` to search}
}
\value{
a list with elements `tests` and `expectation` with character strings of the corresponding test/expectation pairs
}
\description{
This function takes a pair of reults/expectation and performs a comparison.
}
\details{
Call as argument to TestCompareDeparsedList(getTests()), otherwise which argument has to be modified; e.g.: separate line before: which = -1
}
