% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RtestingHelpers.R
\name{installPackageTests}
\alias{installPackageTests}
\title{Prepare test files for an R-package}
\usage{
installPackageTests(packageDir, testPathes, createReference = TRUE)
}
\arguments{
\item{packageDir}{path to R-package folder structure}

\item{testPathes}{character vector with files containing tests of the package}
}
\description{
Takes pathes to test files and installs these tests in an R-package. Also creates required expectation files.
}
