% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\name{inverseOrder_fromOrder}
\alias{inverseOrder_fromOrder}
\title{Calculates inverseOrder, assuming that the argument is already an \code{order}-vector.}
\usage{
inverseOrder_fromOrder(p)
}
\description{
Calculates inverseOrder, assuming that the argument is already an \code{order}-vector.
}
