% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RtestingHelpers.R
\name{LogAt1}
\alias{LogAt1}
\title{Log message at level 1}
\usage{
LogAt1(s)
}
\arguments{
\item{s}{Message to be printed}
}
\value{
invisible message or NULL if nothing was logged
}
\description{
Call function Log with a log level of 1.
}
