% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\name{Rank}
\alias{Rank}
\title{Computes order so that inverseOrder after order is the identity}
\usage{
Rank(p)
}
\description{
Computes order so that inverseOrder after order is the identity
}
\examples{
v = runif(1e2);
print(all(sort(v)[inverseOrder(v)] == v))
}
