% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RtestingHelpers.R
\name{runTestFunction}
\alias{runTestFunction}
\title{Run tests defined in functions}
\usage{
runTestFunction(testName, logger = LogAt1)
}
\arguments{
\item{testName}{charater vector with function names}

\item{logger}{function that prints logging information}
}
\value{
Test result summary is returned as a list of lists with components \code{result}, the boolean test state, \code{NsubTests}, the number of tests performed in the function
}
\description{
Tests are defined by a charachter vector containing function names. Each function is run using \code{runTestFunctionSingle}.
}
