% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RtestingHelpers.R
\name{runTestFunctionSingle}
\alias{runTestFunctionSingle}
\title{Run tests defined in single function}
\usage{
runTestFunctionSingle(testName, logger = LogAt1)
}
\arguments{
\item{testName}{function name as charater vector with one element}

\item{logger}{function that prints logging information}
}
\value{
Test result summary is returned as a list with components \code{result}, the boolean test state, \code{NsubTests}, the number of tests performed in the function
}
\description{
The function is exepected to call \code{TestMe()} to invoke testing. The function is called and test results are collected.
}
