% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RtestingHelpers.R
\name{runTests}
\alias{runTests}
\title{Run tests in isolation}
\usage{
runTests(
  testsFolder = firstDef(options("testme")$testme$testsFolder, "./Rtests"),
  expectationsFolder = firstDef(options("testme")$testme$expectationsFolder,
    "./Rtests/RtestsExpectations"),
  sourceFiles = options("testme")$testme$sourceFiles,
  isolateSession = TRUE,
  useGit = TRUE
)
}
\arguments{
\item{testsFolder}{Folder to be tests. Default folder is taken from `options('testme')$testme$testsFolder`.}

\item{expectationsFolder}{Folder where expectations are stored. Default folder is taken from `options('testme')$testme$expectationsFolder`.}

\item{sourceFiles}{pathes to files that should be sourced prior to running the tests. These would typically hold initialization code.}

\item{isolateSession}{boolean to indicate whether a new R session should be started (using `Rscript`) or whether to run tests in the current R session}

\item{useGit}{logical to indicate whether test vivifications should be commited}
}
\value{
returns 0 on success, value greater 0 if tests failed
}
\description{
Runs tests in a folder in a separate R session.
}
