% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RtestingHelpers.R
\name{runTestsInternal}
\alias{runTestsInternal}
\title{Initialze test environment and run tests}
\usage{
runTestsInternal(
  testsFolder = "Rtests",
  expectationsFolder = "RtestsExpectations",
  useGit = TRUE,
  Ndash = 100,
  logLevel = 4
)
}
\arguments{
\item{testsFolder}{Folder to be tests. Default folder is taken from `options('testme')$testme$testsFolder`.}

\item{expectationsFolder}{Folder where expectations are stored. Default folder is taken from `options('testme')$testme$expectationsFolder`.}
}
\value{
returns 0 on success, value greater 0 if tests failed
}
\description{
Internal function for initialzing test environment and running tests stored in a folder
}
