% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\name{Sprintfl}
\alias{Sprintfl}
\title{Produce string by substituting placeholders}
\usage{
Sprintfl(.fmt, values, sprintf_cartesian = FALSE, envir = parent.frame())
}
\arguments{
\item{s}{template string}

\item{d}{values to substitute into \code{s}}

\item{template}{template for substitution pattern. Within this pattern \code{__DICT_KEY__} is
substituted for a key in \code{d}. This string \code{k} is substituted in \code{s} with \code{d[[k]]}.}
}
\description{
The function behaves similar to sprintf, except that character sequences to be substituted are
indicated by name. To be implemented: *-specifications
}
\examples{
\dontrun{
Sprintf('These are N \%{N} characters.', list(N = 10));
Sprintf('These are N \%{N}d characters.', list(N = 10));
Sprintf('These are N \%{N}02d characters.', list(N = 10));
}
}
