% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RtestingHelpers.R
\name{TestCompareDeparsed}
\alias{TestCompareDeparsed}
\title{Legacy internal test function (TestCompareDeparsed)}
\usage{
TestCompareDeparsed(
  result,
  expectation,
  modes = as.list(rep("compare", length(result)))
)
}
\arguments{
\item{result}{list of results of the current test function}

\item{expectation}{list of Deparsed, expected results}

\item{modes}{comparison modes}
}
\description{
This function takes reult/expectation arguments and performs a comparison.
}
