% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RtestingHelpers.R
\name{TestCompareDeparsedList}
\alias{TestCompareDeparsedList}
\title{Legacy internal test function}
\usage{
TestCompareDeparsedList(pair, modes = as.list(rep("compare", length(result))))
}
\arguments{
\item{modes}{comparison modes}
}
\description{
This function takes a pair of reults/expectation and performs a comparison.
}
