% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testme.h.R
\name{testMe}
\alias{testMe}
\title{Testing}
\usage{
testMe(data, dep = NULL)
}
\arguments{
\item{data}{the data as a data frame}

\item{dep}{a string naming the dependent variable from \code{data},
variable must be numeric}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$info} \tab \tab \tab \tab \tab a table \cr
  \code{results$oneSmartTable} \tab \tab \tab \tab \tab a table \cr
  \code{results$residuals} \tab \tab \tab \tab \tab an output \cr
  \code{results$multi_residuals} \tab \tab \tab \tab \tab an output \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$info$asDF}

\code{as.data.frame(results$info)}
}
\description{
Something here
}
\examples{
\dontrun{
some code here
}
}
