% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RtestingHelpers.R
\name{testmeDir}
\alias{testmeDir}
\title{Run all tests defined in a folder}
\usage{
testmeDir(
  dir = "Rtests",
  expectationsFolder = "Rtests/RtestsExpectations",
  filePattern = ".R$",
  useGit = T,
  logLevel = 4,
  print = T
)
}
\arguments{
\item{dir}{Folder with R-scripts containing tests}
}
\description{
Collect files from a folder using a pattern, extract testing functions and run tests.
}
