% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RtestingHelpers.R
\name{testmeEnvInit}
\alias{testmeEnvInit}
\title{Initialize testing environment}
\usage{
testmeEnvInit(
  expectationsFolder = Sprintf("\%{d}s/RtestsExpectations", d = tempdir()),
  logger = LogAt1
)
}
\arguments{
\item{expectationsFolder}{Path to folder to store test expectations in. Defaults to the \code{RtestsExpectations} sub-folder of argument \code{d}.}

\item{logger}{Function used to print messages. Defaults to \code{LogAt1} which prints to stderr and includes the date}

\item{d}{Base folder holding tests. Defauts to \code{tempdir()}}
}
\value{
Returns the internal environment used by testme
}
\description{
Create package internal environment to hold global state. Calling this function is not required under normal operation. This function is called internally, when whole files or directories are tested. It needs to be called when single testing functions are to be evaluated in an ad-hoc fashion (see vignette).
}
