% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\name{valueCombinations}
\alias{valueCombinations}
\title{Return all value combinations appearing in a data frame}
\usage{
valueCombinations(d)
}
\description{
Return all value combinations appearing in a data frame
}
\examples{
\dontrun{
combs = valueCombinations(iris);
}
}
