% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RtestingHelpers.R
\name{installPackageTests}
\alias{installPackageTests}
\title{Prepare test files for an R-package}
\usage{
installPackageTests(packageDir, testPathes, createReference = TRUE)
}
\description{
Prepare test files for an R-package
}
