% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RtestingHelpers.R
\name{installPackageTests}
\alias{installPackageTests}
\title{Prepare test files for an R-package}
\usage{
installPackageTests(
  packageDir,
  testPathes,
  createReference = TRUE,
  asCran = FALSE
)
}
\arguments{
\item{packageDir}{path to R-package folder structure}

\item{testPathes}{character vector with files containing tests of the package}

\item{createReference}{boolean to indicate whether a reference output is to be created for R when the package tests are run by standard R functions.}

\item{asCran}{boolean to indicate whether tests should be prepared for running on CRAN}
}
\value{
undefined return value.
}
\description{
Takes pathes to test files and installs these tests in an R-package. Also creates required expectation files.
}
