% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RtestingHelpers.R
\name{runPackageTests}
\alias{runPackageTests}
\title{Run tests intalled in R-package}
\usage{
runPackageTests(packageName, useGit = FALSE)
}
\arguments{
\item{packageName}{name of the R-package to be tested}

\item{useGit}{boolean to indicate whether git commit should be ran on the testFolder after running tests}
}
\value{
value of runTests
}
\description{
This is a convenience function to call \code{runTests()} on the test folder of an R-package. The tests are assumed to have been installed with \code{installPackageTests} before.
}
