% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RtestingHelpers.R
\name{TestCompareDeparsed}
\alias{TestCompareDeparsed}
\title{Legacy internal test function (TestCompareDeparsed)}
\usage{
TestCompareDeparsed(
  result,
  expectation,
  modes = as.list(rep("compare", length(result)))
)
}
\arguments{
\item{result}{list of results of the current test function}

\item{expectation}{list of Deparsed, expected results}

\item{modes}{comparison modes}
}
\value{
Returns a list containing the test results (see details)
}
\description{
This function takes reult/expectation arguments and performs a comparison.
}
\details{
The return value is a list with the following components  

\itemize{
	\item{result: }{Logical being true when all tests passed, false otherwise.}
	\item{NsubTests: }{Integer value containing the number of subtests performed.}
}
}
