% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RtestingHelpers.R
\name{TestCompareDeparsedList}
\alias{TestCompareDeparsedList}
\title{Legacy internal test function (TestCompareDeparsedList)}
\usage{
TestCompareDeparsedList(pair, modes = as.list(rep("compare", length(result))))
}
\arguments{
\item{pair}{list with elements `test` (current computation) and `expectation` (deparsed, expected result) to be compared}

\item{modes}{comparison modes}
}
\value{
list with test results (see TestCompareDeparsed)
}
\description{
This function takes a pair of reults/expectation and performs a comparison.
}
\seealso{
{TestCompareDeparsed()} for a definition of the return value
}
