% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RtestingHelpers.R
\name{TestMe}
\alias{TestMe}
\title{Run tests defined in the current function}
\usage{
TestMe(mode = list(), which = -2)
}
\description{
Searches the current function for test definitions, runs the tests and compares to expectations.
}
\examples{
\dontrun{
myTests = function() {
  T1 = 1 + 1;
  TestMe();
}
}
}
