% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RtestingHelpers.R
\name{testmeFile}
\alias{testmeFile}
\title{Run all tests defined in an R-script}
\usage{
testmeFile(file, expectationsFolder, useGit, print = F, logger = LogAt1)
}
\arguments{
\item{file}{R-scripts containing tests}
}
\description{
Collect testing functions from file and run tests
}
